DRTII-1727 Move tests into appropriate classes
